Update Item Display and separate Icon Components - #4639
